+++ _docs_website/components/Doc.js
@@ -1,9 +1,6 @@
 import React from 'react'
 import dynamic from 'next/dynamic'
 import value from '../static/examples.json'
-const mockerDataGenerator = dynamic(import('../../build/main/index.js'), {
-    ssr: false
-})
 const Editor = dynamic(import('./Editor'), { ssr: false })
 const SEditor = dynamic(import('./Split-Editor'), { ssr: false })
 
@@ -23,7 +20,6 @@ export default class Index extends React.Component {
     onChangeCode = ([value]) => {
         this.setState({ value })
         try {
-            let mocker = mockerDataGenerator
             let compiled = eval(value)
             compiled.build(data => {
                 this.setState({